repo cleanup and documentation improvements
