+++ .gitignore
@@ -149,3 +149,6 @@ rsconnect/
 
 # Paper figures
 Analysis/figures*/*
+
+# tmp csv
+data_production/tmp.csv
